+++ libclamav/bytecode_priv.h
@@ -173,7 +173,7 @@ struct cli_bc_ctx {
     uint16_t *opsizes;
     char *values;
     operand_t *operands;
-    uint32_t file_size;
+    size_t file_size;
     int outfd;
     off_t off;
     fmap_t *fmap;
@@ -182,12 +182,12 @@ struct cli_bc_ctx {
     struct cli_bc_hooks hooks;
     struct cli_exe_info exeinfo;
     uint32_t lsigcnt[64];
-    uint32_t lsigoff[64];
+    size_t lsigoff[64];
     uint32_t pdf_nobjs;
     struct pdf_obj **pdf_objs;
     uint32_t *pdf_flags;
-    uint32_t pdf_size;
-    uint32_t pdf_startoff;
+    size_t pdf_size;
+    size_t pdf_startoff;
     unsigned pdf_phase;
     int32_t pdf_dumpedid;
     const struct cli_exe_section *sections;
